issue #106. Moved r427 test.
