+++ src/fparser/two/tests/test_fortran2003.py
@@ -502,40 +502,6 @@ def test_char_length():  # R426
     assert str(obj) == '(:)'
 
 
-def test_char_literal_constant():  # R427
-
-    tcls = Char_Literal_Constant
-    obj = tcls('NIH_"DO"')
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == 'NIH_"DO"'
-    assert repr(obj) == 'Char_Literal_Constant(\'"DO"\', \'NIH\')'
-
-    obj = tcls("'DO'")
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == "'DO'"
-    assert repr(obj) == 'Char_Literal_Constant("\'DO\'", None)'
-
-    obj = tcls("'DON''T'")
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == "'DON''T'"
-
-    obj = tcls('"DON\'T"')
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == '"DON\'T"'
-
-    obj = tcls('""')
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == '""'
-
-    obj = tcls("''")
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == "''"
-
-    obj = tcls('"hey ha(ada)\t"')
-    assert isinstance(obj, tcls), repr(obj)
-    assert str(obj) == '"hey ha(ada)\t"'
-
-
 def test_logical_literal_constant():  # R428
 
     tcls = Logical_Literal_Constant
